+++ apollo-adminservice/src/test/resources/application.properties
@@ -30,3 +30,5 @@ spring.h2.console.enabled = true
 spring.h2.console.settings.web-allow-others=true
 
 spring.main.allow-bean-definition-overriding=true
+
+apollo.audit.log.enabled = true
