+++ apollo-audit/apollo-audit-impl/pom.xml
@@ -46,6 +46,11 @@
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter-data-jpa</artifactId>
     </dependency>
+
+    <dependency>
+      <groupId>com.ctrip.framework.apollo.apollo-audit</groupId>
+      <artifactId>apollo-audit-api</artifactId>
+    </dependency>
   </dependencies>
 
 </project>
