+++ apollo-audit/apollo-audit-impl/src/main/java/com/ctrip/framework/apollo/audit/service/ApolloAuditLogService.java
@@ -52,6 +52,7 @@ public class ApolloAuditLogService {
         .happenTime(new Date())
         .build();
 
+    auditLog.setId(0);
     logRepository.save(auditLog);
   }
