+++ apollo-audit/apollo-audit-spring-boot-starter/src/main/java/com/ctrip/framework/apollo/audit/ApolloAuditProperties.java
@@ -17,9 +17,7 @@
 package com.ctrip.framework.apollo.audit;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Configuration;
 
-@Configuration
 @ConfigurationProperties(prefix = "apollo.audit.log")
 public class ApolloAuditProperties {
