+++ apollo-audit/apollo-audit-spring-boot-starter/src/main/java/com/ctrip/framework/apollo/audit/aop/ApolloAuditSpanAspect.java
@@ -104,6 +104,7 @@ public class ApolloAuditSpanAspect {
       if (auditLog.attachReturnValue()) {
         dataInfluenceList.addAll(toRealList(returnVal));
       }
+
       logService.logSpan(span);
       if (auditLog.autoCollectDataInfluence()) {
         switch (auditLog.type()) {
@@ -121,8 +122,6 @@ public class ApolloAuditSpanAspect {
             break;
         }
       }
-    } catch (Exception e) {
-      e.printStackTrace();
     }
     return returnVal;
   }
