+++ apollo-biz/src/main/java/com/ctrip/framework/apollo/biz/ApolloBizConfig.java
@@ -16,12 +16,9 @@
  */
 package com.ctrip.framework.apollo.biz;
 
-import org.springframework.boot.autoconfigure.AutoConfigurationPackage;
 import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 
 @EnableAutoConfiguration
 @Configuration
