+++ apollo-biz/src/test/resources/application.properties
@@ -28,3 +28,5 @@ spring.jpa.defer-datasource-initialization=true
 
 spring.h2.console.enabled = true
 spring.h2.console.settings.web-allow-others=true
+
+apollo.audit.log.enabled = true
