+++ apollo-configservice/src/main/resources/application.properties
@@ -22,3 +22,5 @@
 
 # You may change the following config to activate different database profiles like h2/postgres
 spring.profiles.group.github = mysql
+
+apollo.audit.log.enabled = true
