+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/controller/ApolloAuditLogController.java
@@ -19,6 +19,7 @@ package com.ctrip.framework.apollo.portal.controller;
 import com.ctrip.framework.apollo.audit.entity.ApolloAuditLog;
 import com.ctrip.framework.apollo.audit.service.ApolloAuditLogService;
 import com.ctrip.framework.apollo.portal.spi.UserInfoHolder;
+
 import java.util.List;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
