+++ apollo-portal/src/test/java/com/ctrip/framework/apollo/openapi/v1/controller/AbstractControllerTest.java
@@ -35,19 +35,18 @@ import org.springframework.web.client.RestTemplate;
 @RunWith(SpringJUnit4ClassRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
 public abstract class AbstractControllerTest {
-  @Autowired
-  private HttpMessageConverters httpMessageConverters;
 
   protected RestTemplate restTemplate = (new TestRestTemplate()).getRestTemplate();
+  @Value("${local.server.port}")
+  protected int port;
+  @Autowired
+  private HttpMessageConverters httpMessageConverters;
 
   @PostConstruct
   protected void postConstruct() {
     restTemplate.setErrorHandler(new DefaultResponseErrorHandler());
     restTemplate.setMessageConverters(httpMessageConverters.getConverters());
   }
-
-  @Value("${local.server.port}")
-  protected int port;
 
   protected String url(String path) {
     return "http://localhost:" + port + path;
