+++ apollo-portal/src/test/resources/application.properties
@@ -31,3 +31,5 @@ spring.h2.console.settings.web-allow-others=true
 
 spring.session.store-type=none
 spring.main.allow-bean-definition-overriding=true
+
+apollo.audit.log.enabled = true
